Improve Github Integration instructions
There were hard to find. I removed the image (it was a bit confusing to see the check-run there, I think there should be an image of the github integration section), but we can always bring it back.

I made the instructions "pop" a bit more. Instead of using a single sentence, it's now an indexed list.